[tcp-echo] Enhancement: simplify metric logging
The current log format makes it harder to parse the output using
automated scripts, and these metrics are not suitable for human
consumption any preprocessing.

Now the header is just printed once, while the line markers are printed
in each line so that grep can be used on the log files easily.

"npushed" is now reset at each log interval, so that we can keep track
of how many packets were tx and rx in each log cycle.