runtime: network: states: remove unneeded string format!() calls
Using format!() for such cases is unnecessary and inefficient. Using
string literals is a better option.

Also, merged arms to avoid duplication, but more importantly to make it
evident that we handle both cases in the same way. This will also
uncover any issues if at all because it draws attention to itself.