ci: workflows: conditional release build to speed up PR merge loop
We decided to run only the debug build for pull request branches. The
debug + release build will run when we merge pull requests into the dev
branch. This will save a lot of PR loop time!

The debug + release build will run when changes are pushed to the
'unstable' and 'main' branches as well.

Given that we have hardly ever encountered issues just in the
release-specific builds, it's unlikely that once the PR branch succeeds
dev CI runs may fail. Of course, it may still happen, but quite low
possibility.

Finally, there is a way to run the release build for PR branches as well
by ending the PR branch name with "do-release" text.