demikernel: use string literals over heap alloc and format!() if possible
Use string literals instead of heap allocations wherever possible. This
will result in better memory management and better performance because
string literals are stored in the read-only binary section and not on
the heap. At runtime, only the reference to the read-only section is
used, which is more efficient.

Avoid bringing in the unnecessary machinary to support format!() if it's
not needed. This results in more readable code as well due to less
visual clutter as well.