bindings: use const for thread_local default init value
By using const for this value, it can be evaluated at compile time. So
no runtime is spent in this allocation. Also, there is no impact on
memory management (i.e. drop()) for this value by making it const.

Regardless, this value is overwritten with a newly initialized LibOS in
code.