ci: code improvements for readability and maintainability
The ci code was littered with unnecessary visual and semantic clutter,
removed that in this first installment of cleanup.

There is still a lot of unnecessary complexity and clutter that should
be cleaned up subsequently. (Hints: too many super() calls are very hard
to reason about and typically point to design issues that should be
addressed soon.)

ci: fix typos